refactor: move table state management logic out of Home
Home was making a lot of mutations to TableData structs that should be
managed by the tables themselves. Logic was moved out of Home's
try_add_extra_reqs() and into table's try_collect_reqs() method.

- Removed unecessarys getters for TableData to get mutable references
  from Home.
- Added more granular count/fetching pattern matching so not every table
  will trigger a refectch when a POST/UPDATE/DELETE happens.
- Table state isn't changed when related requestst are made, rather a
  "future state" is recorded and then state is synced when the table
  handles a sucessful response.

Hopefully this unspaghetti-s the API a little more. At the very least
this refactor allowed the integration tests to be written. There may be
more instances where TableData concerns are not properly separated from
Home.